data_ingest/requirements.txt
asyncpg==0.29.0
openai>=1.3.7
bs4>=0.0.1
typer[all]>=0.9.0
structlog>=24.0.0 
pytest>=7.4.0 
prometheus-client>=0.19.0
tenacity>=8.2.3
datasketch>=1.5.8
sentry-sdk>=1.39.1
feedparser>=6.0.11
aiohttp>=3.9.5 
aiosqlite>=0.19.0 
langdetect>=1.0.9
emoji>=2.10.0
tiktoken>=0.5.1 
PyYAML>=6.0
redis>=5.0.0 